requirements.txt
altair==4.2.2
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
entrypoints==0.4
feedparser==6.0.11
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.3
holidays==0.77
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
nltk==3.9.1
numpy==1.24.3
packaging==25.0
pandas==2.0.3
pillow==11.3.0
plotly==5.15.0
prophet==1.1.5
protobuf==3.20.3
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pywin32==311
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
rich==14.1.0
rpds-py==0.26.0
scikit-learn==1.3.0
scipy==1.15.3
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.19
stanio==0.5.1
streamlit==1.22.0
tenacity==8.5.0
textblob==0.17.1
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
xgboost==1.7.5
zipp==3.23.0
